Correct Makefile
A build id is always included in the final executable in the 'notes' header. If you are not explicit where it should be located in the final binary then, its likely that notes section would come before '.text' section and your multiboot header would be not recognized. Simple check can save hours of debugging!